Refactor: Gamewebsocket의 멤버변수로 MessageManager 선언 후 웹소켓으로 수신한 message를 관리
